Broke frontend into separate components
