+++ public/src/components/instructions.js
@@ -1,0 +1,15 @@
+var React = require('react');
+
+var Instructions = React.createClass({
+  render: function(){
+    return (
+      <div>
+        <h3>
+          Click and hold the microphone button, while pronouncing the word shown below
+        </h3>
+      </div>
+    );
+  }
+});
+
+module.exports = Instructions;
